backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
google-generativeai==0.3.2
pymongo==4.6.1
PyMuPDF==1.24.14
Werkzeug==3.0.1
gunicorn==21.2.0
boto3==1.35.0
PyJWT==2.8.0
